refactor(studio): 收敛共享投影与知识查询边界
