+++ src/studio/view-models/trajectory-evidence.ts
